// src/app/features/complex-form/enums/complex-form-label.enum.ts
export enum ComplexFormLabel {

    FORM_TITLE = 'Inscription',

    SECTION_PERSONAL_INFO_TITLE = 'Informations personnelles',
    INPUT_FIRST_NAME_LABEL = 'Prénom',
    INPUT_LAST_NAME_LABEL = 'Nom',

    SECTION_CONTACT_PREFERENCE_TITLE = 'Comment préférez-vous être contacté(e) ?',

    SECTION_EMAIL_TITLE = 'Email',
    INPUT_EMAIL_LABEL = 'Adresse mail',
    INPUT_CONFIRM_EMAIL_LABEL = 'Confirmer votre adresse mail',

    SECTION_PHONE_TITLE = 'Numéro de téléphone',
    INPUT_PHONE_LABEL = 'Téléphone',

    SECTION_LOGIN_TITLE = 'Informations de connexion',
    INPUT_USERNAME_LABEL = 'Nom d\'utilisateur',
    INPUT_PASSWORD_LABEL = 'Mot de passe',
    INPUT_CONFIRM_PASSWORD_LABEL = 'Confirmer votre mot de passe'

}